hw06/src/main/java/ru/otus/repositories/BookRepositoryJpaImpl.java
package ru.otus.repositories;

import ru.otus.models.Book;
import org.springframework.stereotype.Repository;
import ru.otus.models.Comment;

import javax.persistence.*;
import javax.transaction.Transactional;
import java.util.List;
import java.util.Optional;

@Repository
public class BookRepositoryJpaImpl implements BookRepositoryJpa {

    @PersistenceContext
    private EntityManager em;

    @Transactional
    @Override
    public Book save(Book book) {
        if (book.getId() == 0) {
            em.persist(book);
            return book;
        } else {
            return em.merge(book);
        }
    }

    @Override
    public Optional<Book> findById(long id) {
        return Optional.ofNullable(em.find(Book.class, id));
    }

    @Transactional
    @Override
    public void deleteById(long id) {
        Query query = em.createQuery("delete from Book s where s.id = :id");
        query.setParameter("id", id);
        query.executeUpdate();
    }

    @Override
    public List<Book> findAll() {
        TypedQuery<Book> query = em.createQuery("select b from Book b join fetch b.genre", Book.class);
        return query.getResultList();
    }

    @Override
    public List<Book> findByName(String name) {
        TypedQuery<Book> query = em.createQuery(
                "select s from Book s where s.name = :name", Book.class);
        query.setParameter("name", name);
        return query.getResultList();
    }

    @Transactional
    @Override
    public Comment addComment(Comment comment) {
        if (comment.getId() == 0) {
            em.persist(comment);
            return comment;
        } else {
            return em.merge(comment);
        }
    }

    @Override
    public List<Comment> findAllCommentsByBookId(long bookId) {
        TypedQuery<Comment> query = em.createQuery("select c from Comment c where c.book.id = :book_id", Comment.class);
        query.setParameter("book_id", bookId);
        return query.getResultList();
    }

    @Transactional
    @Override
    public void deleteAllCommentsByBookId(long bookId) {
        Query query = em.createQuery("delete from Comment c where c.book.id = :book_id");
        query.setParameter("book_id", bookId);
        query.executeUpdate();
    }

}
